services/family-api/requirements.txt
# Core Framework
langgraph==0.2.56
langgraph-checkpoint-postgres==2.0.11
langchain-core==0.3.28
langchain-ollama==0.2.2

# API
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.6
pydantic-settings==2.7.1
email-validator==2.2.0

# Database
asyncpg==0.30.0
psycopg2-binary==2.9.10
redis==5.2.1

# Memory & RAG
mem0ai==0.1.39
qdrant-client==1.12.1

# HTTP Clients
httpx==0.28.1
requests==2.32.3

# Task Queue (for future Celery integration)
celery==5.4.0

# Utilities
python-dotenv==1.0.1
python-multipart>=0.0.9
aiofiles>=23.2.020
pydantic-core==2.27.2

# System Monitoring (for dashboard)
psutil==6.1.1
aiohttp==3.11.10

# Observability
prometheus-client==0.21.1
opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-instrumentation-fastapi==0.50b0
opentelemetry-instrumentation-sqlalchemy==0.50b0
opentelemetry-instrumentation-redis==0.50b0
opentelemetry-exporter-otlp==1.29.0
python-json-logger==3.2.1
slowapi==0.1.9  # Rate limiting

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt<5.0  # passlib compatibility - version 5.0+ has breaking changes

# Multimodal Content Processing
aiofiles==24.1.0
Pillow==11.0.0
python-magic==0.4.27
pytesseract==0.3.13
speechrecognition==3.10.4
PyPDF2==3.0.1
python-docx==1.1.2
pydub==0.25.1

# Telegram Integration
python-telegram-bot==21.6

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
httpx==0.28.1  # for test client

# Development
mypy==1.14.1
black==24.10.0
ruff==0.8.5
